requirements.txt
Flask==2.3.3
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
